% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kloss.R
\name{calcAllRates}
\alias{calcAllRates}
\title{Calculate all peptide and protein kloss.}
\usage{
calcAllRates(
  e,
  method = "combined",
  ag.metric = "mean",
  ag.weights = "both",
  in.all = 2
)
}
\arguments{
\item{e}{a pSILAC object.}

\item{method}{the method used to calculate protein-level rates (default 'combined'). See ?calcProteinsKloss}

\item{ag.metric}{the aggregation metric used for protein-level rates (default 'mean'). See ?calcProteinsKloss}

\item{ag.weights}{the method to calculate weights for aggregation (default 'variance'). Ignored if ag.metric != 'mean'. See ?calcProteinsKloss}

\item{in.all}{whether to use only peptides quantified in all samples. See ?calcProteinsKloss}
}
\value{
The updated pSILAC object.
}
\description{
A wrapper that calls, in order, calcRIAkloss(), getHoLkloss() and calcProteinsKloss().
}
