% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHoLkloss.R
\name{calcHoLkloss}
\alias{calcHoLkloss}
\title{Calculate ln(H/L+1)-based k_loss for all peptides}
\usage{
calcHoLkloss(e, ncores = NULL)
}
\arguments{
\item{e}{a pSILAC object.}
}
\value{
The updated pSILAC object.
}
\description{
A wrapper to apply getHoLmod to all peptides of the dataset.
}
