% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHoLkloss.R
\name{calcHoLkloss_v2}
\alias{calcHoLkloss_v2}
\title{Calculate Logarithmic H/L-based k_loss for all Peptides}
\usage{
calcHoLkloss_v2(o, ncores = NULL, tryRobust = FALSE)
}
\arguments{
\item{o}{A `pSILAC` object, which contains the data for peptide H/L ratios and associated metadata.}

\item{ncores}{An integer specifying the number of cores for parallel processing. If `NULL`, it defaults to using all available cores minus one.}

\item{tryRobust}{Logical; if `TRUE`, attempts to fit a robust linear model (requires the `MASS` package). Ignored if fewer than 10 data points are available.}
}
\value{
The updated `pSILAC` object, with a new `hol.kloss` slot containing the calculated k_loss values and associated statistics (standard error, sum of squared residuals, and number of points) for each peptide.
}
\description{
This function calculates the ln(H/L + 1)-based k_loss for each peptide in a `pSILAC` dataset by applying the `getHoLmod` function. 
It handles data across multiple time points and samples, with optional parallel processing to speed up computations.
}
\details{
The function iterates over each unique sample in the dataset and calculates k_loss values for each peptide, either in parallel or sequentially. 
For each sample, the calculated metrics include the k_loss, its standard error, sum of squared residuals, and the number of points used in the calculation.
A summary message is displayed for each sample, indicating the number of successfully calculated k_loss values and any missing values.
}
