% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcKdeg.R
\name{calcKdeg}
\alias{calcKdeg}
\title{Calculate Protein Degradation Rates}
\usage{
calcKdeg(o, rate_df = NULL, type = "kperc", perc_neg = 0.01)
}
\arguments{
\item{o}{A pSILAC object containing protein degradation data and a `protein.kloss` matrix.
The object must include `protein.kloss` values, which are required for degradation rate calculation.}

\item{rate_df}{An optional data frame containing custom degradation rate values (`kcd`) for each sample.
This data frame should contain columns:
- `sample`: Sample names that must match with those in the pSILAC object.
- `kcd`: The custom degradation rate values for each sample.
Required only when `type = "kcd"`.}

\item{type}{A character string specifying the rate calculation type.
Options are:
- `kperc`: Estimate theoretical kcd values based on the specified quantile (`perc_neg`).
- `kcd`: Use custom kcd values provided in `rate_df`.
Default is `kperc`.}

\item{perc_neg}{A numeric value specifying the quantile to use when estimating theoretical kcd values if `type = "kperc"`.
Default is `0.01`, representing the 1st percentile.}
}
\value{
A modified pSILAC object with an added `protein.kdeg` matrix, representing the calculated
degradation rates for each protein. Any kdeg values <= 0 are replaced by NA before storing.
}
\description{
This function calculates protein degradation rates based on the protein-level
`kloss` values from a pSILAC object. It supports two types of calculations:
estimating theoretical degradation rates using a specified quantile (`kperc`)
or using custom rates provided in a data frame (`kcd`).
}
