% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kloss.R
\name{calcNLIkloss}
\alias{calcNLIkloss}
\title{Calculate k_loss for all peptides on the basis of normalized light intensities}
\usage{
calcNLIkloss(e, startIntensity = "median", ncores = NULL)
}
\arguments{
\item{e}{a pSILAC object.}

\item{startIntensity}{character; either "model" (fit the model with an intersect), "max" (use maximum H+L intensitiy as intersect), or "median" (use median of non-null H+L intensities). Default "median".}

\item{ncores}{the number of cores to use (defaults to detected cores minus 1)}
}
\value{
The updated pSILAC object.
}
\description{
A wrapper to apply getNLImod to all peptides of the dataset.
}
