% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kloss_peptide.R
\name{calcRIAkloss}
\alias{calcRIAkloss}
\title{Calculate RIA-based k_loss for all peptides}
\usage{
calcRIAkloss(e, ncores = NULL)
}
\arguments{
\item{e}{a pSILAC object.}

\item{ncores}{the number of cores to use (defaults to detected cores minus 1)}
}
\value{
The updated pSILAC object.
}
\description{
A wrapper to apply getRIAmod to all peptides of the dataset.
}
