% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kloss.R
\name{compareParametersResults}
\alias{compareParametersResults}
\title{compareParametersResults}
\usage{
compareParametersResults(ll, replicates)
}
\arguments{
\item{ll}{a list, as produced by the compareParameters function.}

\item{replicates}{A vector of length equal to the number of samples, indicating which samples are replicates of the same condition. 
If this is contained in the design data.frame of a pSILAC object, you may retrieve it through: o$design[which(!duplicated(o$design$sample)),"replicateColumnNames"]}
}
\value{
Nothing.
}
\description{
Plots the results of the 'compareParameters' function, taking only the proteins which are given a value in all analyses.
}
