% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatepSILACObject.R
\name{.generatepSILACObject_v0}
\alias{.generatepSILACObject_v0}
\title{Generate pSILAC class object, outdated}
\usage{
.generatepSILACObject_v0(
  dataset,
  design,
  inputDataType = "spectronaut",
  requant = "remove",
  aggregate.replicates = NA,
  filterPeptides = T,
  ncores = 1,
  imputeMethod = "normD1",
  noiseCutoff = 8
)
}
\arguments{
\item{dataset}{A tab-separated file or `data.frame` in R containing precursor-level data.

If provided as a file, `dataset` must include:
- Columns for light and heavy precursor intensities
- A unique precursor ID column
- A protein ID column

If provided as a `data.frame` in R, `dataset` must:
- Use precursor IDs as row names
- Include columns for light and heavy intensities}

\item{design}{A tab-separated file or `data.frame` in R defining the experimental design.

- **File format**: The first column must be the raw file name. The file should also contain:
  - A `sample` column for sample identifiers
  - A `time` column in numeric format (hours)
  - Optional columns:
    - `replicate` column (numeric), for replicate IDs if a replicate design analysis is desired
    - `color` column to customize data plotting
- For replicate samples to be averaged, they must have the same value in the `sample` column.

- **Data.frame format**: If loaded as a `data.frame` in R, `design` must use raw file names as row names.
- **Template**: Use `generate_design_template()` to create a customizable design table template.}

\item{inputDataType}{A character string specifying the type of input data. Options include `spectronaut`, `diann`, `fragpipe`, `maxquant`, or `openswath`.}

\item{requant}{Relevant for OpenSwath data only. Specifies handling of requantified values with a score greater than `0.05`.
- Options:
  - `keep`: Retain requantified values
  - `remove`: Exclude requantified values (default)
  - `impute`: Replace requantified values with imputed values
- **Data requirements**: Assumes columns are prepended with `Intensity_` and `score_` for each sample/timepoint in `dataset`.}

\item{aggregate.replicates}{Method for aggregating replicate values:
- Options: `median` (recommended) or `mean`. Set to `NA` if no replicates are present or should not be aggregated.
- For replicate averaging, all replicates should have the same value in the `sample` column of `design`.}

\item{filterPeptides}{Logical; if `TRUE`, removes peptides without any lysine (K) or arginine (R) residues from the analysis (recommended).}

\item{ncores}{Integer; the number of cores to use for parallel computation. Defaults to `1`.}

\item{noiseCutoff}{Numeric; specifies a cutoff for light and heavy intensities. If `NULL`, no cutoff is applied.}
}
\value{
a pSILAC object.
}
\description{
Creates a pSILAC object. Needs a design table and precursor-level report from an DIA-MS data processing software such as
Spectronaut, DIA-NN, and Fragpipe. 
These can be provided as path or as data loaded in R environment.
}
\keyword{internal}
