% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kloss.R
\name{.medianNonNull}
\alias{.medianNonNull}
\title{Calculate the Median of Non-Null, Non-Zero Values}
\usage{
.medianNonNull(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector from which the median of non-zero values is to be calculated.}

\item{na.rm}{A logical value indicating whether \code{NA} values should be stripped before 
the computation proceeds. Default is \code{TRUE}.}
}
\value{
The median of the non-zero values in \code{x}. If no values greater than zero are
        found, the function returns \code{NA}.
}
\description{
This function computes the median of a numeric vector, ignoring values 
that are less than or equal to zero. It returns \code{NA} if no values
greater than zero are present in the vector.
}
\details{
This function first removes all values from the input vector \code{x} that are 
         less than or equal to zero. It then calculates the median of the remaining 
         values. If the resulting vector is empty, \code{NA} is returned.
}
\examples{
# Example 1: Basic usage
x <- c(0, 2, 5, NA, -1, 3)
.medianNonNull(x)

# Example 2: Vector with only negative and zero values
y <- c(-1, 0, -5, NA)
.medianNonNull(y)

}
