% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kloss.R
\name{.modProtPre}
\alias{.modProtPre}
\title{Prepare petides for protein-level kloss calculation}
\usage{
.modProtPre(e, in.all, removeOutliers)
}
\arguments{
\item{e}{a matrix/data.frame of peptides values pSILAC object.}

\item{in.all}{Starting with how many peptides should peptides not quantified in all samples be ignored (default 2), or FALSE to use all peptides, even if not quantified in all samples.}

\item{removeOutliers}{Starting with how many peptides should outlier peptides be removed, or FALSE to disable outlier removal (default 5). See ?removeOutlierPeptides}
}
\value{
A filtered version of e.
}
\description{
Prepare petides for protein-level kloss calculation
}
