% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelProteinsKloss.R
\name{.mwm}
\alias{.mwm}
\title{Calculate Modified Weighted Mean}
\usage{
.mwm(v, w, minVals = 3, na.rm = TRUE)
}
\arguments{
\item{v}{A numeric vector of values for which the weighted mean is to be calculated.}

\item{w}{A numeric vector of weights corresponding to the values in `v`. Must be the same length as `v`.
Missing values in `w` are set to zero.}

\item{minVals}{An integer specifying the minimum number of non-zero weights required to calculate the weighted mean.
If the count of non-zero weights is below this threshold, the function returns the arithmetic mean of `v`.
Default is 3.}

\item{na.rm}{Logical, indicating whether NA values should be removed from `v` and `w` before calculating the mean.
Default is `TRUE`.}
}
\value{
A numeric value representing either the weighted mean of `v` or the arithmetic mean of `v` 
  (if there are fewer than `minVals` non-zero weights).
}
\description{
This function calculates a weighted mean of the values in `v` using weights in `w`. If the number of non-zero weights 
is below a specified threshold (`minVals`), the function returns the arithmetic mean of `v` instead. 
Missing values in `w` are treated as zero.
}
\details{
This function provides flexibility when calculating a weighted mean, returning an unweighted mean if there are 
insufficient non-zero weights. This is useful in cases where an adequate number of reliable weights is required 
to apply weighting. Missing values in the weight vector `w` are automatically set to zero.
}
