% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_diann}
\alias{example_diann}
\title{Example pSILAC dataset analyzed in DIA-NN 1.9}
\format{
An object of class \code{data.frame} with 19999 rows and 58 columns.
}
\usage{
example_diann
}
\description{
This dataset contains a data.frame containing the first 20,000 unique precursors
from the A2780Cis and A2780 parental cell line dataset. This dataset contains 3 dish (biological) replicates
and 4 labeling time points of 1, 4, 12, and 24 hours. 
The data were analyzed and exported from DIA-NN 1.9.
}
\keyword{datasets}
