% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_kcd}
\alias{example_kcd}
\title{Example Kcd file}
\format{
An object of class \code{data.frame} with 6 rows and 2 columns.
}
\usage{
example_kcd
}
\description{
This file contains Kcd values of the A2780 and A2780Cis cell lines.
}
\keyword{datasets}
