% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_maxquant}
\alias{example_maxquant}
\title{Example pSILAC dataset analyzed in MaxQuant}
\format{
An object of class \code{data.frame} with 5000 rows and 224 columns.
}
\usage{
example_maxquant
}
\description{
This dataset contains a data.frame containing the first 5,000 unique peptides
from a publicly available dataset (PXD057850) documented in the study of Frankenfield et al., MCP, 2025
Benchmarking SILAC Proteomics Workflows and Data Analysis Platforms, PMID: 40315959 
Result folder: MaxQuant_dSILAC_CurveFitting.zip, peptides.txt
Contains one condition, measured in 4 time points and four replicates.
}
\keyword{datasets}
