% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_maxquant_design}
\alias{example_maxquant_design}
\title{Example design table for a dataset analyzed in MaxQuant}
\format{
An object of class \code{data.frame} with 16 rows and 4 columns.
}
\usage{
example_maxquant_design
}
\description{
This dataset contains and example design table to run a pSILAC data analysis with replicate aggregation.
Data from publicly available dataset (PXD057850) documented in the study of Frankenfield et al., MCP, 2025
Benchmarking SILAC Proteomics Workflows and Data Analysis Platforms, PMID: 40315959 
Result folder: MaxQuant_dSILAC_CurveFitting.zip, peptides.txt
#' Contains one condition, measured in 4 time points and four replicates.
}
\keyword{datasets}
