% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_pSILAC_object}
\alias{example_pSILAC_object}
\title{Example pSILAC object}
\format{
A \code{pSILAC} class object with fitted turnover parameters and
associated quality-control metrics.
}
\usage{
example_pSILAC_object
}
\description{
An example \code{pSILAC} object, containing precursor- and protein-level turnover rate estimates.
}
\keyword{datasets}
