% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_spectronaut}
\alias{example_spectronaut}
\title{Example pSILAC dataset analyzed in Spectronaut 19}
\format{
An object of class \code{data.frame} with 20000 rows and 56 columns.
}
\usage{
example_spectronaut
}
\description{
This dataset contains a data.frame containing the first 20,000 unique precursors
from the A2780Cis and A2780 parental cell line dataset. This dataset contains 3 dish (biological) replicates
and 4 labeling time points of 1, 4, 12, and 24 hours. 
The data were analyzed and exported from Spectronaut 19.
}
\keyword{datasets}
