% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_spectronaut_isw}
\alias{example_spectronaut_isw}
\title{Example pSILAC dataset analyzed in Spectronaut, Inverted-Spike In workflow}
\format{
An object of class \code{data.frame} with 10000 rows and 80 columns.
}
\usage{
example_spectronaut_isw
}
\description{
This dataset contains a data.frame containing the first 10,000 unique precursors
from the A2780Cis and A2780 parental cell line dataset. This dataset contains 3 dish (biological) replicates
and 4 labeling time points of 1, 4, 12, and 24 hours. 
The data were analyzed and exported from Spectronaut 19.
}
\keyword{datasets}
