% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_spectronaut_pSILAC_object}
\alias{example_spectronaut_pSILAC_object}
\title{Example Output from the demo analysis}
\format{
An object of class \code{pSILAC} of length 18.
}
\usage{
example_spectronaut_pSILAC_object
}
\description{
This file is a pSILAc class object with all precursor and protein-level rates estimated.
}
\keyword{datasets}
