% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMonotoneTimePoint1.R
\name{filter_monotone_trend_tp1only}
\alias{filter_monotone_trend_tp1only}
\title{Filter the first time point based on monotone filtering}
\usage{
filter_monotone_trend_tp1only(data, mode = NULL)
}
\arguments{
\item{data}{A data frame containing values from one pSILAC labeling series.}

\item{mode}{Specifies whether the filtering is performed with the "RIA", "HOL", or "NLI" values.}
}
\value{
A filtered data frame.
}
\description{
The input is a data frame  frame split into a list based on grouping defined in the design object.
}
