% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterLinearRegression.R
\name{filterLinearRegression}
\alias{filterLinearRegression}
\title{Filter Data Based on Linear Regression of ln(H/L + 1) Values}
\usage{
filterLinearRegression(
  o,
  skip_time_point = 1,
  R2_cutoff = 0.9,
  p_cutoff = 0.05
)
}
\arguments{
\item{o}{A `pSILAC` object containing the experimental data, specifically the `hol` and `ria` data frames.}

\item{skip_time_point}{Integer. Specifies how many early time points should be excluded from the linear regression analysis.}

\item{R2_cutoff}{R2 cutoff to remove the outliers (0.9 by default), only curves below this cutoff will be filtered}

\item{p_cutoff}{Grubb's test P value cutoff to remove the outliers (0.05 by default)
Must be 0 or 1. Default is 1.}
}
\value{
A modified `pSILAC` object with filtered data. The `hol` and `ria` data frames are updated by removing rows 
(based on specific criteria in the linear regression results), and an additional `test.outliers.hol` list contains 
the regression model fits for each sample.
}
\description{
Applies linear regression filtering to the ln(H/L + 1) data in a `pSILAC` object.
This function performs linear regression across each labeling series in the dataset, removes specific 
data points based on goodness-of-fit and outlier thresholds, and returns an updated `pSILAC` object.
}
\details{
This function performs the following steps:
\itemize{
  \item Fits a linear model to each series (excluding an initial time point if specified) and stores regression metrics.
  \item Identifies data points where the full model R-squared is below 0.9 and the p-value from Grubbs' test (outlier test) is less than 0.05.
  \item Sets the selected data points to `NA` in both the `hol` and `ria` data frames and recombines the series into their original format.
}
}
\examples{
# Assuming `pSILAC_obj` is a valid pSILAC object with appropriate data
filtered_pSILAC <- filterLinearRegression(pSILAC_obj, skip_time_point = 1)

}
