% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMonotone.R
\name{filterMonotone}
\alias{filterMonotone}
\title{A wrapper function to filter data based on monotone filtering}
\usage{
filterMonotone(o, skip_time_point = 1)
}
\arguments{
\item{o}{a pSILAC object}

\item{skip_time_point}{Specifies how many early time points should be excluded from valid value filtering. Default is 1.}
}
\value{
A filtered data frame.
}
\description{
A wrapper to apply the filter_monotone_trend_tp1excluded function to all pSILAC labeling series in the dataset.
}
