% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMonotone.R
\name{filterMonotone_skipTimePoints}
\alias{filterMonotone_skipTimePoints}
\title{Filter data based on monotone filtering}
\usage{
filterMonotone_skipTimePoints(
  data,
  values_cutoff = 2,
  skip_time_point = 1,
  mode = NULL
)
}
\arguments{
\item{data}{A data frame containing values from one pSILAC labeling series.}

\item{values_cutoff}{Specifies the minimum number of valid values to keep per pSILAC labeling series. Default is 2.}

\item{skip_time_point}{Specifies how many early time points should be excluded from valid value filtering. Default is 1.}

\item{mode}{Specifies whether the filtering is performed with the "RIA", "HOL", or "NLI" values.}
}
\value{
A filtered data frame.
}
\description{
The input is a data frame  frame split into a list based on grouping defined in the design object. Do everything with HOL, and then filter ther RIA data frane based on it?
}
