% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMonotone.R
\name{filterMonotoneSkipTimePoints}
\alias{filterMonotoneSkipTimePoints}
\title{Filter data based on monotone filtering}
\usage{
filterMonotoneSkipTimePoints(data, skip_time_point = 1, mode = NULL)
}
\arguments{
\item{data}{A data frame containing values from one pSILAC labeling series.}

\item{skip_time_point}{Specifies how many early time points should be excluded from valid value filtering. Default is 1.}

\item{mode}{Specifies whether the filtering is performed with the "RIA", "HOL", or "NLI" values.}
}
\value{
A filtered data frame.
}
\description{
The input is a data frame  frame split into a list based on grouping defined in the design object.
}
