% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMonotoneTp1Only.R
\name{filterMonotoneTp1Only}
\alias{filterMonotoneTp1Only}
\title{Filter First Time Point Based on Monotone Trend}
\usage{
filterMonotoneTp1Only(data, mode = NULL)
}
\arguments{
\item{data}{A data frame representing values from a single pSILAC labeling series, where each row corresponds to one observation.}

\item{mode}{Specifies the type of data values:
- `"RIA"`: Assumes a monotone decreasing trend for the first time point.
- `"HOL"`: Assumes a monotone increasing trend for the first time point.
- `"NLI"`: Assumes a monotone decreasing trend for the first time point.
The mode determines the filtering criteria applied.}
}
\value{
Returns a filtered data frame with the same columns as the input `data`, where the first time point is adjusted based on the specified monotonic trend criteria.
}
\description{
This function applies monotone trend filtering to the first time point in a pSILAC labeling series, based on the specified mode.
Depending on whether the values represent `RIA`, `HOL`, or `NLI` data, the function filters rows according to the monotonicity assumption:
a decreasing trend for `RIA` and `NLI`, and an increasing trend for `HOL`.
}
