% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterValidValues.R
\name{filterValidValues_skipTimePoints}
\alias{filterValidValues_skipTimePoints}
\title{Filter data based on valid values, skipping early time points}
\usage{
filterValidValues_skipTimePoints(data, values_cutoff = 2, skip_time_point = 1)
}
\arguments{
\item{data}{A data frame containing values from one pSILAC labeling series.}

\item{values_cutoff}{Specifies the minimum number of valid values to keep per pSILAC labeling series. Default is 2.}

\item{skip_time_point}{Specifies how many early time points should be excluded from valid value filtering. Default is 1.}
}
\value{
a filtered data frame.
}
\description{
The input is a data frame  frame split into a list based on grouping defined in the design object.
}
