% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLinearModel.R
\name{fitLinearModel}
\alias{fitLinearModel}
\title{Filter Data Based on Monotone Linear Regression}
\usage{
fitLinearModel(data, skip_time_point = 1, time_points = time_points)
}
\arguments{
\item{data}{A data frame where each row represents a pSILAC labeling series, with columns 
representing time points.}

\item{skip_time_point}{Integer specifying the number of early time points to exclude from 
the analysis. Must be 0 or 1; default is 1.}

\item{time_points}{A numeric vector indicating the time points corresponding to each column 
in `data`. The length must match the number of columns in `data`.}
}
\value{
A data frame with the following columns:
\describe{
  \item{R2}{R-squared of the reduced model, excluding the first time point if skipped.}
  \item{intercept}{Intercept of the reduced model.}
  \item{slope}{Slope of the reduced model.}
  \item{grubbs_pval}{P-value from Grubbs' test for the residual at the first time point, indicating 
                     potential outliers if it has the maximum residual.}
  \item{time_point1}{Logical, indicating if time point 1 was included in the full model.}
  \item{R2_full}{R-squared of the full model, including all time points if time point 1 has data.}
  \item{intercept_full}{Intercept of the full model.}
  \item{slope_full}{Slope of the full model.}
}
}
\description{
This function fits a linear regression model to each row in a pSILAC dataset, 
optionally excluding early time points. The function filters data based on specified criteria, 
including a minimum number of valid values and trends in the data. It returns model summary 
statistics, including R-squared values, slopes, intercepts, and p-values from Grubbs' test to 
identify outliers.
}
\details{
This function iteratively fits a linear model to each row in `data`, focusing on the `ln(H/L + 1)` 
transformation of pSILAC values. Two models are fit:
- **Reduced model**: Excludes the first time point if `skip_time_point` is set to 1.
- **Full model**: Includes all time points, if the first time point has a valid value.

After fitting, each row is evaluated based on:
- R-squared values from both reduced and full models (`R2` and `R2_full`).
- Estimated intercept and slope for both models.
- Grubbs' test p-value (`grubbs_pval`), which identifies outliers at the first time point 
  if it has the largest residual in absolute terms.
}
