% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{gdensity}
\alias{gdensity}
\title{Plot multiple density distributions}
\usage{
gdensity(
  thelist,
  cols = "black",
  xlab = "Value",
  ylab = "Density",
  xlim = 0,
  lwd = 2,
  main = "",
  ...
)
}
\arguments{
\item{thelist}{a list or data.frame containing only number values.}

\item{cols}{a vector of colors, or length=length(thelist)}

\item{xlab}{passed to the plot function.}

\item{ylab}{passed to the plot function.}

\item{xlim}{Either a single value (max x), or a vector of length 2 passed to the plot function.}

\item{lwd}{line width, passed to the plot function.}

\item{main}{the plot title (default empty)}

\item{...}{addtional arguments passed to the plot function.}
}
\value{
Plots the density distributions.

Plots the density distributions.
}
\description{
Maps colors onto numeric values.
}
\examples{
gdensity(list(A=1:3,B=2:4,C=3:5),c("red","green","blue"))


Plot multiple density distributions

gdensity(list(A=1:3,B=2:4,C=3:5),c("red","green","blue"))

}
