% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDesignTemplate.R
\name{generateDesignTemplate}
\alias{generateDesignTemplate}
\title{Generate design table template}
\usage{
generateDesignTemplate(dataset, inputDataType = "spectronaut")
}
\arguments{
\item{dataset}{path to the data exported from MS data processing file}

\item{inputDataType}{can be "spectronaut", "diann", or "fragpipe", output from other software are not supported (for this function)}
}
\description{
Generate design table template
}
