% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{getGlobalVar}
\alias{getGlobalVar}
\title{getGlobalVar}
\usage{
getGlobalVar(y, x)
}
\arguments{
\item{y}{A numeric vector containing the values}

\item{x}{A numeric vector of same length as y, containing the groups to which the corresponding value in y belongs.}
}
\value{
Returns the group-adjusted variance.
}
\description{
Returns the global variance (removing differences between groups)
}
