% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNLImod.R
\name{getNLImod}
\alias{getNLImod}
\title{Calculate the Rate of Isotope Loss (k_loss) Based on Normalized Light Intensities (NLI)}
\usage{
getNLImod(x, y, y0 = NULL, a = 0.1)
}
\arguments{
\item{x}{A numeric vector of timepoints associated with the NLI measurements.}

\item{y}{A numeric vector of normalized light intensity values at each timepoint in `x`. Must be the same length as `x`.}

\item{y0}{The intercept of the model, representing the initial intensity. If `NULL` (default), it will be estimated during model fitting.}

\item{a}{A numeric value providing the initial guess for `k_loss` during model fitting. Default is 0.1.}
}
\value{
A numeric vector of length 4 containing the following elements:
\itemize{
  \item \code{k_loss}: The estimated rate of loss for the light isotope.
  \item \code{SE_k_loss}: The standard error of the estimated \code{k_loss}.
  \item \code{residual_sum}: The sum of residuals from the fitted model, indicating the goodness of fit.
  \item \code{n_points}: The number of data points used in the model, excluding any missing values in `y`.
}
}
\description{
This function models the rate of loss of a light isotope (`k_loss`) by fitting a non-linear model 
to normalized light intensity (NLI) values across specified timepoints. The model assumes exponential 
decay of NLI over time, allowing estimation of `k_loss`.
}
\details{
This function applies a non-linear least squares method to fit an exponential decay model to the NLI values across timepoints.
If there is only one valid timepoint, a simplified calculation of `k_loss` is provided, using the intercept if available.
If model fitting fails, `NA` values are returned for `k_loss` and `SE_k_loss`, while `n_points` indicates the number of valid points.
}
