% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{getNormValues}
\alias{getNormValues}
\title{getNormValues}
\usage{
getNormValues(m, dsd, nb)
}
\arguments{
\item{m}{desired mean}

\item{dsd}{desired standard deviation}

\item{nb}{number of values}
}
\value{
Returns a numeric vector of (deterministic) values with a specific mean and SD.
}
\description{
Get a set of values with a specific (approximative) mean and SD.
}
