% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pSILAC.utils.R
\name{getPeptides}
\alias{getPeptides}
\title{getPeptides}
\usage{
getPeptides(x, peptides = NULL, protein = NULL, returnValues = NA, search = T)
}
\arguments{
\item{x}{a pSILAC object.}

\item{peptides}{a list of peptide names/sequences.}

\item{protein}{a protein name.}

\item{returnValues}{values to return. Either 'light' (light channel intensities), 'heavy', 'RIA', 'hol' (log-transformed heavy/light), 'RIA.kloss', 'NLI.kloss' or 'hol.kloss'.
If NA (default), will return the peptides (from x$peptides).}

\item{sample}{the sample(s) for which to fetch values. NULL (default) fetches all samples. Ignored if returnValues is NA, 'RIA.kloss' or 'hol.kloss'.}
}
\value{
data.frame.
}
\description{
Returns given peptides, or peptides corresponding to a given protein. Only one of 'peptides' or 'protein' should be set.
}
