% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRIAmod.R
\name{getRIAmod}
\alias{getRIAmod}
\title{Calculate the Rate of Isotope Loss (k_loss) Based on Relative Isotope Abundance (RIA)}
\usage{
getRIAmod(x, y, a = 0.1)
}
\arguments{
\item{x}{A numeric vector of timepoints corresponding to the measurements of RIA.}

\item{y}{A numeric vector of RIA values associated with each timepoint in `x`. Must be the same length as `x`.}

\item{a}{A numeric value specifying the initial guess for `k_loss` during model fitting. Default is 0.1.}
}
\value{
A numeric vector of length 4 containing the following elements:
\itemize{
  \item \code{k_loss}: The estimated rate of loss for the light isotope.
  \item \code{SE_k_loss}: The standard error of the estimated \code{k_loss}.
  \item \code{residual_sum}: The sum of residuals from the fitted model, indicating goodness of fit.
  \item \code{n_points}: The number of data points used in the model, excluding any missing values in `y`.
}
}
\description{
This function models the rate of loss for a light isotope (`k_loss`) using a non-linear model based on 
the relative isotope abundance (RIA) values across specified timepoints. The model assumes an exponential decay 
of RIA over time, allowing estimation of `k_loss`.
}
\details{
This function uses non-linear least squares to fit an exponential decay model to the RIA values over time.
If only one valid timepoint is available, a simplified calculation is used. If the model fitting fails,
`NA` values are returned for \code{k_loss} and \code{SE_k_loss}, while the number of valid points (`n_points`) is provided.
}
