% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pSILAC.class.R
\name{head.pSILAC}
\alias{head.pSILAC}
\title{summary method for pSILAC objects.}
\usage{
\method{head}{pSILAC}(x)
}
\arguments{
\item{x}{a pSILAC object.}
}
\value{
the head() of each attribute of x.
}
\description{
summary method for pSILAC objects.
}
