% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{imputeAll}
\alias{imputeAll}
\title{imputeAll}
\usage{
imputeAll(e, x = NULL, method = "normD", sd.offset = 1, log.transform = T)
}
\arguments{
\item{e}{The numeric matrix.}

\item{x}{The groups to which each column in e belongs.}

\item{method}{The imputation method; either: "norm", "normD", or "zero2min",}

\item{sd.offset}{The number of standard deviation to substract from the mean.}

\item{log.transform}{Logical; whether to log-transform for the purpose of calculating the normal distribution.}
}
\value{
Returns a numeric vector of (deterministic) values with a specific mean and SD.
}
\description{
Impute all missing values in a matrix or data.frame.
}
