% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{kloss.boxplot}
\alias{kloss.boxplot}
\title{Plots a boxplot of a protein's peptide-level kloss values.}
\usage{
kloss.boxplot(
  object,
  protein,
  method = "RIA",
  main = NULL,
  in.all = NULL,
  removeOutliers = NULL,
  normalize = F
)
}
\arguments{
\item{object}{A pSILAC object.}

\item{protein}{A protein name (i.e. uniprot ID)}

\item{method}{The method from which kloss values should be taken. Either 'RIA' or 'hol'.}

\item{main}{The plot's title, passed to the plot function. Defaults to the protein's name/id.}

\item{in.all}{Starting with how many peptides should peptides not quantified in all samples be ignored (default 2), or FALSE to use all peptides, even if not quantified in all samples.}

\item{removeOutliers}{Starting with how many peptides should outlier peptides be removed, or FALSE to disable outlier removal (default 5). See ?removeOutlierPeptides}
}
\value{
Nothing, but plots a figure.
}
\description{
Plots a boxplot of a protein's peptide-level kloss values.
}
