% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{kloss.heatmap}
\alias{kloss.heatmap}
\title{Plots a heatmap of a protein's peptide-level kloss values.}
\usage{
kloss.heatmap(object, protein, method = "RIA")
}
\arguments{
\item{object}{A pSILAC object.}

\item{protein}{A protein name (i.e. uniprot ID)}

\item{method}{The method from which kloss values should be taken. Either 'RIA' or 'hol'.}
}
\value{
Nothing, but plots a figure.
}
\description{
Plots a heatmap of a protein's peptide-level kloss values.
}
