% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{maketrans}
\alias{maketrans}
\title{Make transparent}
\usage{
maketrans(tcol, alpha = 100)
}
\arguments{
\item{alpha}{a numeric value between 0 and 255 indicating the degree of transparency.}

\item{x}{a color specification (see \code{\link[grDevices]{col2rgb}}}
}
\value{
A color code.
}
\description{
Adds transparency to a color.
}
\examples{
maketrans("blue")

}
