% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kloss.R
\name{MdRE}
\alias{MdRE}
\title{Calculates median relatibe error (MdRE)}
\usage{
MdRE(x, y = NULL)
}
\arguments{
\item{x}{Either a numeric vector, or a numeric matrix/data.frame with samples as columns, and features as rows.}

\item{y}{Should be NULL if x is a matrix/data.frame, or (if x is a vector) a numeric vector of length equal to that of x.}
}
\value{
Either the median relative error (if x and y are vectors), or a symmetric matrix (of size=ncol(x)) of pair-wise MdRE if x is a matrix/data.frame.
}
\description{
Calculates median relatibe error (MdRE)
}
