% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pSILAC.utils.R
\name{medianNorm}
\alias{medianNorm}
\title{Normalize Data by Median Scaling}
\usage{
medianNorm(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric matrix or data frame to be normalized.}

\item{na.rm}{Logical; if `TRUE`, NA values are removed when calculating medians. Default is `TRUE`.}
}
\value{
A data frame with normalized values, where each column is scaled by its median to ensure comparability across samples.
}
\description{
This function normalizes columns of a matrix or data frame by scaling each column to its median.
}
\details{
The function computes a size factor for each column based on the median value, then scales each value by these factors.
This normalization method is often used to adjust for variations in sample size or sequencing depth.
}
