% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeLightChannel.R
\name{normalizeLightChannel}
\alias{normalizeLightChannel}
\title{normalizeLightChannel}
\usage{
normalizeLightChannel(o, method = "linear", timepoints = NULL, removeNAs = T)
}
\arguments{
\item{o}{a pSILAC object.}

\item{method}{normalization method, either 'linear' (default) or 'geometric'.}

\item{timepoints}{an optional vector indicating the timepoint for each sample. If NULL (default), falls back to the value of the `time` column of the design.}

\item{removeNAs}{logical; whether to remove NA values before normalizing (default TRUE)}
}
\value{
data.frame.
}
\description{
Normalizes the light channel so that the channel sum is constant. 
The normalized light channel can be used for the NLI method.
}
