% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plColorMap}
\alias{plColorMap}
\title{plColorMap}
\usage{
plColorMap(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
A vector of colors of the same length as x.
}
\description{
Maps colors onto numeric values.
}
\examples{
plColorMap(1:3)

}
