% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotKlossDensities}
\alias{plotKlossDensities}
\title{Plots each sample's distribution of protein kloss rates.}
\usage{
plotKlossDensities(
  o,
  remove.NAs = T,
  xlim = c(0, 0.1),
  main = "Protein-level rates of loss",
  xlab = "K_loss"
)
}
\arguments{
\item{o}{A pSILAC object.}

\item{remove.NAs}{Whether to remove proteins whose kloss is missing in any of the samples. Otherwise, all proteins will be used for the samples in which they are present.}

\item{xlim}{Either a single value (max x), or a vector of length 2 passed to the 'gdensity' function. Default c(0,0.1)}

\item{main}{Plot title, passed to the plot function.}

\item{xlab}{X-axis label, passed to the plot function.}
}
\value{
Nothing, but plots a figure.
}
\description{
Plots each sample's distribution of protein kloss rates.
}
