% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotMissingValues}
\alias{plotMissingValues}
\title{Histogram of missing values per peptide}
\usage{
plotMissingValues(x, channel = "RIA", ...)
}
\arguments{
\item{x}{A pSILAC object.}

\item{channel}{The channel in which to look for missing values. Either 'heavy', 'light', 'RIA' or 'hol' (H/L). 
RIA and hol should be equivalent, since both are missing when either the heavy or the light is missing.}

\item{a}{The starting value of the k_loss for model-fitting.}
}
\value{
A histogram.
}
\description{
Plots the distribution of missing values per peptide in the desired channel/value.
}
