% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotProtein}
\alias{plotProtein}
\title{plotProtein}
\usage{
plotProtein(o, protein, removeOutliers = 5)
}
\arguments{
\item{o}{A pSILAC object.}

\item{protein}{A protein name (i.e. uniprot ID)}

\item{removeOutliers}{Starting with how many peptides should outlier peptides be removed, or FALSE to disable outlier removal (default 5). See ?removeOutlierPeptides}
}
\value{
Nothing, but plots a figure.
}
\description{
Creates a layout of 4 plots of a protein's degradation: 
 - the protein's relative isotope abundance over time (calls 'plotProteinRIA'), 
 - a boxplot of the protein's peptide-level RIA-based kloss (calls 'kloss.boxplot'), 
 - the protein's ln(H/L+1) ratio over time (calls 'plotProteinHol'), 
 - a boxplot of the protein's peptide-level H/L-based kloss (calls 'kloss.boxplot'),
}
