% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotProteinRIA}
\alias{plotProteinRIA}
\title{plotProteinRIA}
\usage{
plotProteinRIA(
  object,
  protein,
  samples = NULL,
  peptides.alpha = 150,
  plot.legend = T,
  main = NULL,
  xlab = "Time (hours)",
  ...
)
}
\arguments{
\item{object}{A pSILAC object.}

\item{protein}{A protein name (i.e. uniprot ID)}

\item{samples}{The samples to plot. If NULL, all samples are plotted.}

\item{peptides.alpha}{The alpha value of individual data points.}

\item{plot.legend}{Whether to plot a figure legend (default TRUE).}

\item{main}{The plot's title, passed to the plot function. Defaults to the protein's name/id.}

\item{xlab}{Passed to the plot function. Defaults to 'Time (hours)'.}

\item{...}{additional arguments passed to the plot function.}
}
\value{
Nothing, but plots a figure.
}
\description{
Plots the relative isoform abundance for a given protein over time.
}
