% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pSILAC.class.R
\name{print.pSILAC}
\alias{print.pSILAC}
\title{print method for pSILAC objects.}
\usage{
\method{print}{pSILAC}(x)
}
\arguments{
\item{x}{a pSILAC object.}
}
\value{
nothing, but prints info on the object...
}
\description{
print method for pSILAC objects.
}
