% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeOutlierPeptides.R
\name{removeOutlierPeptides}
\alias{removeOutlierPeptides}
\title{Remove Outlier peptides}
\usage{
removeOutlierPeptides(x, min.size = 3, pthreshold = 1e-04)
}
\arguments{
\item{x}{a numeric matrix or data.frame, with peptides as rows.}

\item{min.size}{the minimum number of rows in x before outlier removal is performed.}

\item{pthreshold}{the probability threshold above which values are removed.}
}
\value{
an object of the same class and ncol as x.
}
\description{
Removes rows whose median is has a probability greater than ptreshold in the empirical frequency distribution of all rows' median. 
Not recommended for less than 5 rows.
}
