% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pSILAC.class.R
\name{summary.pSILAC}
\alias{summary.pSILAC}
\title{summary method for pSILAC objects.}
\usage{
\method{summary}{pSILAC}(x)
}
\arguments{
\item{x}{a pSILAC object.}
}
\value{
nothing, but prints info on the object...
}
\description{
summary method for pSILAC objects.
}
