% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kloss.R
\name{aggregateKloss}
\alias{aggregateKloss}
\title{aggregateKloss}
\usage{
aggregateKloss(
  e,
  samples,
  ag.metric = "mean",
  ag.weights = "variance",
  unique.weights = T,
  in.all = 2,
  removeOutliers = 5,
  returnSD = F
)
}
\arguments{
\item{e}{a data.frame of peptide-level kloss, as in the RIA.kloss or hol.kloss (or NLI.kloss) slots of a pSILAC object.}

\item{samples}{the vector of samples' names.}

\item{ag.metric}{the aggregation metric used for protein-level rates (default 'mean'). Either 'mean' or 'median'. If 'mean' is chosen, and if ag.weights is not null, a weighted mean will be performed.}

\item{ag.weights}{the method to calculate weights for weighted mean (default 'variance'). Ignored if ag.metric != 'mean'. Either 'variance' (1/variance of the model's fit), 'nbpoints' (number of datapoints used for the fit), or 'both' (nbpoints/variance).}

\item{unique.weights}{whether unique weights should be used for each peptides (default T), or if false a weight is calculated individually for each sample.}

\item{in.all}{Starting with how many peptides should peptides not quantified in all samples be ignored (default 2), or FALSE to use all peptides, even if not quantified in all samples.}

\item{removeOutliers}{Starting with how many peptides should outlier peptides be removed, or FALSE to disable outlier removal (default 5). See ?removeOutlierPeptides}

\item{returnSD}{Logical; whether to return also the SD of the kloss (default FALSE).}
}
\value{
A vector of length=length(samples) with protein-level kloss corresponding to each sample.
}
\description{
Aggregate peptides' kloss into a protein-level kloss. Rather than calling this function directly, use calcProteinsKloss
}
