% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNLIkloss.R
\name{calcNLIkloss}
\alias{calcNLIkloss}
\title{Calculate k_loss for All Peptides Based on Normalized Light Intensities}
\usage{
calcNLIkloss(o, startIntensity = "max", ncores = NULL)
}
\arguments{
\item{o}{A `pSILAC` object containing the data to be analyzed.}

\item{startIntensity}{Character. Specifies the method for determining the starting intensity for model fitting:
\itemize{
  \item \code{"model"} - Fits the model with an intersect.
  \item \code{"max"} - Uses the maximum H+L intensity as the intersect (default).
  \item \code{"median"} - Uses the median of non-null H+L intensities as the intersect.
}}

\item{ncores}{Integer. The number of cores to use for parallel computation. If \code{NULL} (default), the function 
detects available cores and uses all but one.}
}
\value{
The input `pSILAC` object with an added `NLI.kloss` element, containing the calculated \( k_{loss} \), 
standard error, sum of squared residuals, and the number of points used in the fitting for each sample.
}
\description{
This function applies the `getNLImod` function to all peptides in a pSILAC dataset to calculate the decay rate 
constant \( k_{loss} \) for normalized light intensities. It uses a specified starting intensity as the intersect for fitting 
the model, and supports parallel computation to speed up processing.
}
\details{
This function performs the following steps:
\itemize{
  \item Uses parallel processing, if multiple cores are available, to apply `getNLImod` on each peptide across different samples.
  \item Calculates \( k_{loss} \), the standard error, and other fit metrics for each peptide, based on the specified starting intensity method.
}
}
