% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kloss.R
\name{calcProteinsKloss}
\alias{calcProteinsKloss}
\title{Calculate all protein kloss.}
\usage{
calcProteinsKloss(
  x,
  method = "combined",
  ag.metric = "mean",
  ag.weights = "both",
  unique.weights = T,
  in.all = 2,
  removeOutliers = 5,
  ncores = NULL,
  tryRobust = F,
  returnKlossTableOnly = F,
  returnSD = F
)
}
\arguments{
\item{x}{a pSILAC object.}

\item{method}{the method used to calculate protein-level rates (default 'combined'); either 'RIA', 'hol', 'NLI', or 'combined' (uses both RIA- and NLI-based peptide kloss).}

\item{ag.metric}{the aggregation metric used for protein-level rates (default 'mean'). Either 'mean', 'median' or 'remodel'. If 'mean' is chosen, and if ag.weights is not null, a weighted mean will be performed.}

\item{ag.weights}{the method to calculate weights for weighted mean (default 'variance'). Ignored if ag.metric != 'mean'. Either 'variance' (1/variance of the model's fit), 'nbpoints' (number of datapoints used for the fit), or 'both' (nbpoints/variance).}

\item{unique.weights}{whether unique weights should be used for each peptides (default T), or if false a weight is calculated individually for each sample.}

\item{in.all}{Starting with how many peptides should peptides not quantified in all samples be ignored (default 2), or FALSE to use all peptides, even if not quantified in all samples.}

\item{removeOutliers}{Starting with how many peptides should outlier peptides be removed, or FALSE to disable outlier removal (default 5). See ?removeOutlierPeptides}

\item{ncores}{Number of cores to use (only for remodeling). Defaults to all cores minus 1.}

\item{tryRobust}{Whether to try fitting a robust linear model when remodeling on the basis of H/L ratio (requires the 'MASS' library). Disabled by default.}

\item{returnKlossTableOnly}{Logical; whether to return only the kloss table, rather than the whole pSILAC object (default).}

\item{returnSD}{Logical; whether to return also the SD of the kloss (default FALSE).}
}
\value{
The updated pSILAC object, or the protein kloss table if returnKlossTableOnly is TRUE.
}
\description{
Aggregates peptide-level kloss for all proteins. This requires that calcRIAkloss, calcHoLkloss, and/or calcNLIkloss be run first.
}
