% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRIAkloss.R
\name{calcRIAkloss}
\alias{calcRIAkloss}
\title{Calculate RIA-based k_loss for All Peptides in a Dataset}
\usage{
calcRIAkloss(o, ncores = NULL)
}
\arguments{
\item{o}{A `pSILAC` object containing the dataset to be processed. 
This object should include RIA data and experimental design information.}

\item{ncores}{The number of cores to use for parallel processing. 
Defaults to one less than the total number of detected cores on the system.
If set to 1 or NULL, single-core processing is used.}
}
\value{
The updated `pSILAC` object with the calculated `k_loss` values added in the 
        `RIA.kloss` field. This field includes `k_loss`, standard error, sum of squared residuals, 
        and the number of data points used per sample.
}
\description{
This function serves as a wrapper to apply `getRIAmod` to each peptide in a pSILAC dataset, 
calculating RIA-based (Relative Isotopic Abundance) `k_loss` values. The results are stored 
in the pSILAC object's `RIA.kloss` field. Multi-core processing is supported for faster computation.
}
\details{
The function iterates over each unique sample in the dataset, applying `getRIAmod` 
to calculate the RIA-based `k_loss` for each peptide. Results are stored in a matrix 
with columns representing `k_loss`, standard error (`kloss.stderr`), sum of squared residuals (`kloss.SSR`), 
and the number of points used (`nbpoints`). If parallel processing is enabled, `parApply` 
is used for faster computation.
}
