% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelProteinsKloss.R
\name{.modProtPre}
\alias{.modProtPre}
\title{Prepare Peptides for Protein-Level k_loss Calculation}
\usage{
.modProtPre(o, in.all = 2, removeOutliers = 5)
}
\arguments{
\item{o}{A matrix or data frame of peptide values, typically from a pSILAC object.}

\item{in.all}{Numeric or logical, specifying the threshold for peptides not quantified in all samples.
If a numeric value (default is 2), only peptides quantified in `in.all` or more samples are retained. 
Set to `FALSE` to include all peptides, regardless of quantification across samples.}

\item{removeOutliers}{Numeric or logical, specifying the threshold for outlier removal.
If a numeric value (default is 5), removes outlier peptides starting at the specified threshold.
Set to `FALSE` to disable outlier removal. See `?removeOutlierPeptides` for details on outlier identification.}
}
\value{
A filtered version of `o` with peptides potentially removed based on the `in.all` and `removeOutliers` criteria.
}
\description{
This function filters a matrix or data frame of peptide values in preparation for protein-level k_loss calculation. 
It allows for the optional exclusion of peptides not quantified in all samples and removal of outlier peptides.
}
\details{
The function applies two main filtering steps:
\enumerate{
  \item If `in.all` is specified as a numeric value, peptides not quantified in all samples will be ignored if their 
  missing values exceed the threshold.
  \item If `removeOutliers` is specified as a numeric value, outlier peptides are removed using `removeOutlierPeptides` 
  when the peptide count meets the threshold.
}
The function returns the input object `o` with the applied filters or in its original form if filtering criteria are not met.
}
