% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_diann_design}
\alias{example_diann_design}
\title{Example design table for the A2780 pSILAC dataset (analyzed in DIA-NN 1.9)}
\format{
An object of class \code{data.frame} with 24 rows and 5 columns.
}
\usage{
example_diann_design
}
\description{
This dataset contains and example design table to run a pSILAC data analysis without replicate design.
}
\keyword{datasets}
