% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_kcd}
\alias{example_kcd}
\title{Example Kcd dataset}
\format{
A data frame containing Kcd estimates suitable for downstream analysis
and visualization.
}
\usage{
example_kcd
}
\description{
An example dataset containing Kcd values estimated for the A2780 parental
and A2780Cis cell lines.
}
\keyword{datasets}
