% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_spectronaut_design}
\alias{example_spectronaut_design}
\title{Example deign table for the A2780 pSILAC dataset (analyzed in Spectronaut 19)}
\format{
An object of class \code{data.frame} with 24 rows and 5 columns.
}
\usage{
example_spectronaut_design
}
\description{
This dataset contains and example desigb table to run a pSILAC data analysis without replicate design.
}
\keyword{datasets}
