% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMonotoneSkipTimePoints.R
\name{filterMonotoneSkipTimePoints}
\alias{filterMonotoneSkipTimePoints}
\title{Filter Data Based on Monotonic Trend with Skip Time Points}
\usage{
filterMonotoneSkipTimePoints(data, skip_time_point = 1, mode = NULL)
}
\arguments{
\item{data}{A data frame containing values for one pSILAC labeling series. Each row should represent an individual observation, and each column represents a time point.}

\item{skip_time_point}{Integer. Specifies the number of early time points to exclude from the filtering process. Default is 1.}

\item{mode}{Character. Specifies the type of monotonic trend filtering to apply:
\itemize{
  \item \code{"RIA"} - filters for a monotonic decrease.
  \item \code{"NLI"} - filters for a monotonic decrease.
  \item \code{"HOL"} - filters for a monotonic increase.
}}
}
\value{
A filtered data frame with rows that meet the specified monotonic trend criteria. Columns `valid_values`, `increasing_trend`, and `decreasing_trend` are removed from the output.
}
\description{
This function filters a single pSILAC labeling series by applying a monotonic trend 
filter to data columns, with an option to exclude a specified number of early time points from the filtering process.
The input data frame is split into subsets based on the number of valid values in each row.
}
