% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMonotoneTimePoint1.R
\name{filterMonotoneTimePoint1}
\alias{filterMonotoneTimePoint1}
\title{Monotone Filtering for the First Time Point in pSILAC Labeling Series}
\usage{
filterMonotoneTimePoint1(o)
}
\arguments{
\item{o}{a pSILAC object}
}
\value{
A filtered data frame.
}
\description{
This function is a wrapper that applies monotone trend filtering to the first time point in each pSILAC labeling series 
within a dataset. It processes three data frames from a pSILAC object (`RIA`, `HOL`, and `NLI`) and ensures consistent 
column ordering after filtering.
}
