% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterValidValues.R
\name{filterValidValues}
\alias{filterValidValues}
\title{Filter data based on valid values}
\usage{
filterValidValues(o, values_cutoff = 2, skip_time_point = 1)
}
\arguments{
\item{o}{a pSILAC object.}

\item{values_cutoff}{Specifies the minimum number of valid values to keep per pSILAC labeling series (default is 2).}

\item{skip_time_point}{Specifies how many early time points should be excluded from valid value filtering (default is 1).}
}
\value{
The updated pSILAC object with filtered data.
}
\description{
A wrapper to apply the filter_valid_skip_early_tp function to all pSILAC labeling series in the dataset.
}
