% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_design_template.R
\name{generate_design_template}
\alias{generate_design_template}
\title{Generate design table template}
\usage{
generate_design_template(dataset, inputDataType = "spectronaut")
}
\arguments{
\item{dataset}{path to the data exported from MS data processing file}

\item{inputDataType}{can be "spectronaut", "diann", or "fragpipe"}
}
\description{
Generate design table template
}
