% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kloss.R
\name{getHoLmod}
\alias{getHoLmod}
\title{Calculate H/L-based k_loss}
\usage{
getHoLmod(x, y, tryRobust = F)
}
\arguments{
\item{x}{The timepoints.}

\item{y}{The ln(H/L+1) values corresponding to the timepoints.}

\item{tryRobust}{Whether to try to fit a robust linear model (requires the 'MASS' library). Ignored with less than 10 data points.}
}
\value{
A numeric vector of length 4 containing: 
    the k_loss, i.e. ln(H/L+1)slope, 
    the standard error of the slope, 
    the sum of residuals of the linear model, 
    the number of points on which the model was built.
}
\description{
Models the rate of loss of the light isotope (k_loss) on the basis of the slope of the log-transformed isotope ratio.
}
