% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kloss_fitting_functions.R
\name{getNLImod}
\alias{getNLImod}
\title{Calculate NLI-based k_loss}
\usage{
getNLImod(x, y, y0 = NULL, a = 0.1)
}
\arguments{
\item{x}{The timepoints.}

\item{y}{The normalized intensities values corresponding to the timepoints.}

\item{y0}{The intersect of the model. If NULL (default), will attempt to fit it.}

\item{a}{The starting value of the k_loss for model-fitting.}
}
\value{
A numeric vector of length 4 containing: 
    the k_loss,
    the standard error of the k_loss, 
    the sum of residuals of the linear model, 
    the number of points on which the model was built.
}
\description{
Models the rate of loss of the light isotope (k_loss) on the basis of the normalized light intensities
}
