% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kloss.R
\name{getRIAmod}
\alias{getRIAmod}
\title{Calculate RIA-based k_loss}
\usage{
getRIAmod(x, y, a = 0.1)
}
\arguments{
\item{x}{The timepoints.}

\item{y}{The RIA values corresponding to the timepoints.}

\item{a}{The starting value of the k_loss for model-fitting.}
}
\value{
A numeric vector of length 4 containing: 
    the k_loss,
    the standard error of the k_loss, 
    the sum of residuals of the linear model, 
    the number of points on which the model was built.
}
\description{
Models the rate of loss of the light isotope (k_loss) on the basis of the relative isotope abundance (RIA).
}
