% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kloss.R
\name{modelProteinsKloss}
\alias{modelProteinsKloss}
\title{Calculate all protein kloss by fitting all the proteins' peptides on a single model}
\usage{
modelProteinsKloss(
  o,
  method = "combined",
  ag.weights = "variance",
  unique.weights = T,
  in.all = 2,
  removeOutliers = 5,
  ncores = NULL,
  tryRobust = F,
  returnKlossTableOnly = F,
  freeIntersect = F,
  returnSD = F
)
}
\arguments{
\item{o}{a pSILAC object.}

\item{method}{the method used to calculate protein-level rates (default 'combined'); either 'RIA', 'hol' or 'combined' (uses both RIA- and H/L-based peptide kloss).}

\item{ag.weights}{the method to calculate weights for weighted mean (default 'variance'). Ignored if ag.metric != 'mean'. Either 'variance' (1/variance of the model's fit), 'nbpoints' (number of datapoints used for the fit), or 'both' (nbpoints/variance).}

\item{unique.weights}{whether unique weights should be used for each peptides (default T), or if false a weight is calculated individually for each sample.}

\item{in.all}{Starting with how many peptides should peptides not quantified in all samples be ignored (default 2), or FALSE to use all peptides, even if not quantified in all samples.}

\item{removeOutliers}{Starting with how many peptides should outlier peptides be removed, or FALSE to disable outlier removal (default 5). See ?removeOutlierPeptides}

\item{ncores}{Number of cores to use (only for remodeling). Defaults to all cores minus 1.}

\item{tryRobust}{Whether to try fitting a robust linear model when remodeling on the basis of H/L ratio (requires the 'MASS' library). Disabled by default.}

\item{returnKlossTableOnly}{Logical; whether to return only the kloss table, rather than the whole pSILAC object (default).}

\item{freeIntersect}{Logical; whether to fit also the intersect for the NLI-based method.}

\item{returnSD}{Logical; whether to return also the SD of the kloss (default FALSE).}
}
\value{
The updated pSILAC object, or the protein kloss table if returnKlossTableOnly is TRUE.
}
\description{
Aggregates peptide-level kloss for all proteins. This requires that calcRIAkloss and/or calcHoLkloss be run first.
}
