% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pSILAC.R
\name{pSILAC}
\alias{pSILAC}
\title{Generate pSILAC class object}
\usage{
pSILAC(
  dataset,
  design,
  inputDataType = "spectronaut",
  requant = "remove",
  aggregate.replicates = NA,
  filterPeptides = T,
  ncores = 1,
  imputeMethod = "normD1",
  noiseCutoff = 8
)
}
\arguments{
\item{dataset}{a data.frame containing peptides (both heavy and light) as row.names and samples and timepoints as columns (with the addition of a 'Protein' column).}

\item{design}{a data.frame where row.names correspond (optionally with "Intensity_" and "score_" prepended) to columns of 'dataset'. 
In addition, the 'design' object should have columns named 'sample' (character) and 'time' (numeric).}

\item{inputDataType}{can be "spectronaut", "diann", "fragpipe", "maxquant", or "openswath}

\item{requant}{whether to 'keep', 'remove' or 'impute' the requantified values with a score greater than 0.05 (default 'remove', or keep if ). 
This requires that there be in 'dataset' columns prepended with 'Intensity_' and 'score_' for each sample/timepoint.
This option was developed for the output of OpenSwath.}

\item{aggregate.replicates}{Function with which to aggregate replicates ('median' or 'mean'; median recommended), or NA if replicates are absent or should not be aggregated.
If there are replicates, they should have the same value in the column 'sample' of the design data.frame.}

\item{filterPeptides}{If TRUE, removes peptides without any K or R}

\item{ncores}{the number of cores to use (defaults to 1)}

\item{noiseCutoff}{if not NULL, the cutoff will be applied to light and heavy intensities}
}
\value{
a pSILAC object.
}
\description{
Creates a pSILAC object.
}
